# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Management API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type BookingPackageType =
  (typeof BookingPackageType)[keyof typeof BookingPackageType];

export const BookingPackageType = {
  daily: "daily",
  short_stay: "short_stay",
} as const;

export type BookingPaymentMethod =
  (typeof BookingPaymentMethod)[keyof typeof BookingPaymentMethod];

export const BookingPaymentMethod = {
  prepaid: "prepaid",
  pay_at_counter: "pay_at_counter",
} as const;

export type BookingStatus = (typeof BookingStatus)[keyof typeof BookingStatus];

export const BookingStatus = {
  pending: "pending",
  paid: "paid",
  checked_out: "checked_out",
} as const;

export interface Booking {
  id: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  roomId: number;
  guestName: string;
  /** @minimum 1 */
  numGuests: number;
  checkInAt: string;
  checkOutAt: string;
  packageType: BookingPackageType;
  paymentMethod: BookingPaymentMethod;
  status: BookingStatus;
  basePrice: number;
  overtimeFee: number;
  extraBedFee: number;
  totalPrice: number;
  overtimeHours: number;
  notes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type CreateBookingBodyPackageType =
  (typeof CreateBookingBodyPackageType)[keyof typeof CreateBookingBodyPackageType];

export const CreateBookingBodyPackageType = {
  daily: "daily",
  short_stay: "short_stay",
} as const;

export type CreateBookingBodyPaymentMethod =
  (typeof CreateBookingBodyPaymentMethod)[keyof typeof CreateBookingBodyPaymentMethod];

export const CreateBookingBodyPaymentMethod = {
  prepaid: "prepaid",
  pay_at_counter: "pay_at_counter",
} as const;

export interface CreateBookingBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  roomId: number;
  /** @minLength 1 */
  guestName: string;
  /** @minimum 1 */
  numGuests?: number;
  checkInAt: string;
  checkOutAt: string;
  packageType: CreateBookingBodyPackageType;
  paymentMethod: CreateBookingBodyPaymentMethod;
  notes?: string | null;
}

export type UpdateBookingBodyStatus =
  (typeof UpdateBookingBodyStatus)[keyof typeof UpdateBookingBodyStatus];

export const UpdateBookingBodyStatus = {
  pending: "pending",
  paid: "paid",
  checked_out: "checked_out",
} as const;

export interface UpdateBookingBody {
  status?: UpdateBookingBodyStatus;
  checkOutAt?: string;
  notes?: string | null;
  /** @minimum 1 */
  numGuests?: number;
}

export interface BookingsSummary {
  totalBookings: number;
  pendingCount: number;
  paidCount: number;
  checkedOutCount: number;
  totalRevenue: number;
  todayBookings: number;
  todayRevenue: number;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type ListBookingsParams = {
  roomId?: number;
  startDate?: string;
  endDate?: string;
};