# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Booking Management API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: bookings
    description: Booking operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      summary: List all bookings
      parameters:
        - name: roomId
          in: query
          required: false
          schema:
            type: integer
        - name: startDate
          in: query
          required: false
          schema:
            type: string
        - name: endDate
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of bookings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Booking"
    post:
      operationId: createBooking
      tags: [bookings]
      summary: Create a new booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookingBody"
      responses:
        "201":
          description: Created booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "400":
          description: Validation error or conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Booking conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      summary: Get a single booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBooking
      tags: [bookings]
      summary: Update a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBookingBody"
      responses:
        "200":
          description: Updated booking
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Booking"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBooking
      tags: [bookings]
      summary: Delete a booking
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /bookings/summary:
    get:
      operationId: getBookingsSummary
      tags: [bookings]
      summary: Get bookings summary statistics
      responses:
        "200":
          description: Summary statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BookingsSummary"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Booking:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
          minimum: 1
          maximum: 5
        guestName:
          type: string
        numGuests:
          type: integer
          minimum: 1
        checkInAt:
          type: string
          format: date-time
        checkOutAt:
          type: string
          format: date-time
        packageType:
          type: string
          enum: [daily, short_stay]
        paymentMethod:
          type: string
          enum: [prepaid, pay_at_counter]
        status:
          type: string
          enum: [pending, paid, checked_out]
        basePrice:
          type: number
        overtimeFee:
          type: number
        extraBedFee:
          type: number
        totalPrice:
          type: number
        overtimeHours:
          type: number
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - roomId
        - guestName
        - numGuests
        - checkInAt
        - checkOutAt
        - packageType
        - paymentMethod
        - status
        - basePrice
        - overtimeFee
        - extraBedFee
        - totalPrice
        - overtimeHours
        - createdAt
        - updatedAt
    CreateBookingBody:
      type: object
      properties:
        roomId:
          type: integer
          minimum: 1
          maximum: 5
        guestName:
          type: string
          minLength: 1
        numGuests:
          type: integer
          minimum: 1
          default: 1
        checkInAt:
          type: string
          format: date-time
        checkOutAt:
          type: string
          format: date-time
        packageType:
          type: string
          enum: [daily, short_stay]
        paymentMethod:
          type: string
          enum: [prepaid, pay_at_counter]
        notes:
          type: string
          nullable: true
      required:
        - roomId
        - guestName
        - checkInAt
        - checkOutAt
        - packageType
        - paymentMethod
    UpdateBookingBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, paid, checked_out]
        checkOutAt:
          type: string
          format: date-time
        notes:
          type: string
          nullable: true
        numGuests:
          type: integer
          minimum: 1
    BookingsSummary:
      type: object
      properties:
        totalBookings:
          type: integer
        pendingCount:
          type: integer
        paidCount:
          type: integer
        checkedOutCount:
          type: integer
        totalRevenue:
          type: number
        todayBookings:
          type: integer
        todayRevenue:
          type: number
      required:
        - totalBookings
        - pendingCount
        - paidCount
        - checkedOutCount
        - totalRevenue
        - todayBookings
        - todayRevenue
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message