# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Management API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all bookings
 */
export const ListBookingsQueryParams = zod.object({
  roomId: zod.coerce.number().optional(),
  startDate: zod.coerce.string().optional(),
  endDate: zod.coerce.string().optional(),
});

export const listBookingsResponseRoomIdMax = 5;

export const ListBookingsResponseItem = zod.object({
  id: zod.number(),
  roomId: zod.number().min(1).max(listBookingsResponseRoomIdMax),
  guestName: zod.string(),
  numGuests: zod.number().min(1),
  checkInAt: zod.coerce.date(),
  checkOutAt: zod.coerce.date(),
  packageType: zod.enum(["daily", "short_stay"]),
  paymentMethod: zod.enum(["prepaid", "pay_at_counter"]),
  status: zod.enum(["pending", "paid", "checked_out"]),
  basePrice: zod.number(),
  overtimeFee: zod.number(),
  extraBedFee: zod.number(),
  totalPrice: zod.number(),
  overtimeHours: zod.number(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});
export const ListBookingsResponse = zod.array(ListBookingsResponseItem);

/**
 * @summary Create a new booking
 */
export const createBookingBodyRoomIdMax = 5;

export const createBookingBodyNumGuestsDefault = 1;

export const CreateBookingBody = zod.object({
  roomId: zod.number().min(1).max(createBookingBodyRoomIdMax),
  guestName: zod.string().min(1),
  numGuests: zod.number().min(1).default(createBookingBodyNumGuestsDefault),
  checkInAt: zod.coerce.date(),
  checkOutAt: zod.coerce.date(),
  packageType: zod.enum(["daily", "short_stay"]),
  paymentMethod: zod.enum(["prepaid", "pay_at_counter"]),
  notes: zod.string().nullish(),
});

/**
 * @summary Get a single booking
 */
export const GetBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const getBookingResponseRoomIdMax = 5;

export const GetBookingResponse = zod.object({
  id: zod.number(),
  roomId: zod.number().min(1).max(getBookingResponseRoomIdMax),
  guestName: zod.string(),
  numGuests: zod.number().min(1),
  checkInAt: zod.coerce.date(),
  checkOutAt: zod.coerce.date(),
  packageType: zod.enum(["daily", "short_stay"]),
  paymentMethod: zod.enum(["prepaid", "pay_at_counter"]),
  status: zod.enum(["pending", "paid", "checked_out"]),
  basePrice: zod.number(),
  overtimeFee: zod.number(),
  extraBedFee: zod.number(),
  totalPrice: zod.number(),
  overtimeHours: zod.number(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Update a booking
 */
export const UpdateBookingParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateBookingBody = zod.object({
  status: zod.enum(["pending", "paid", "checked_out"]).optional(),
  checkOutAt: zod.coerce.date().optional(),
  notes: zod.string().nullish(),
  numGuests: zod.number().min(1).optional(),
});

export const updateBookingResponseRoomIdMax = 5;

export const UpdateBookingResponse = zod.object({
  id: zod.number(),
  roomId: zod.number().min(1).max(updateBookingResponseRoomIdMax),
  guestName: zod.string(),
  numGuests: zod.number().min(1),
  checkInAt: zod.coerce.date(),
  checkOutAt: zod.coerce.date(),
  packageType: zod.enum(["daily", "short_stay"]),
  paymentMethod: zod.enum(["prepaid", "pay_at_counter"]),
  status: zod.enum(["pending", "paid", "checked_out"]),
  basePrice: zod.number(),
  overtimeFee: zod.number(),
  extraBedFee: zod.number(),
  totalPrice: zod.number(),
  overtimeHours: zod.number(),
  notes: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  updatedAt: zod.coerce.date(),
});

/**
 * @summary Delete a booking
 */
export const DeleteBookingParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get bookings summary statistics
 */
export const GetBookingsSummaryResponse = zod.object({
  totalBookings: zod.number(),
  pendingCount: zod.number(),
  paidCount: zod.number(),
  checkedOutCount: zod.number(),
  totalRevenue: zod.number(),
  todayBookings: zod.number(),
  todayRevenue: zod.number(),
});