# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { BookingPackageType } from "./bookingPackageType";
import type { BookingPaymentMethod } from "./bookingPaymentMethod";
import type { BookingStatus } from "./bookingStatus";

export interface Booking {
  id: number;
  /**
   * @minimum 1
   * @maximum 5
   */
  roomId: number;
  guestName: string;
  /** @minimum 1 */
  numGuests: number;
  checkInAt: Date;
  checkOutAt: Date;
  packageType: BookingPackageType;
  paymentMethod: BookingPaymentMethod;
  status: BookingStatus;
  basePrice: number;
  overtimeFee: number;
  extraBedFee: number;
  totalPrice: number;
  overtimeHours: number;
  notes?: string | null;
  createdAt: Date;
  updatedAt: Date;
}