# lib/api-zod/src/generated/types/createBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookingBodyPackageType } from "./createBookingBodyPackageType";
import type { CreateBookingBodyPaymentMethod } from "./createBookingBodyPaymentMethod";

export interface CreateBookingBody {
  /**
   * @minimum 1
   * @maximum 5
   */
  roomId: number;
  /** @minLength 1 */
  guestName: string;
  /** @minimum 1 */
  numGuests?: number;
  checkInAt: Date;
  checkOutAt: Date;
  packageType: CreateBookingBodyPackageType;
  paymentMethod: CreateBookingBodyPaymentMethod;
  notes?: string | null;
}