# lib/api-zod/src/generated/types/updateBookingBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Booking Management API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateBookingBodyStatus } from "./updateBookingBodyStatus";

export interface UpdateBookingBody {
  status?: UpdateBookingBodyStatus;
  checkOutAt?: Date;
  notes?: string | null;
  /** @minimum 1 */
  numGuests?: number;
}